+++ src/common/components/Navigation.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from "react";
+import React from "react";
 import { style } from "styles";
 import { useNavigate } from "react-router-dom";
