+++ src/common/components/chart/ProgressBarSemiCircle.jsx
