+++ src/common/components/provider/GlobalVariableProvider.jsx
