+++ src/common/utils/constants/styleVariables.js
