+++ src/features/board/components/BoardDetailComment.jsx
