+++ src/features/board/components/BoardDetailInput.jsx
