+++ src/features/board/components/CashDetailModal.jsx
@@ -1,13 +1,14 @@
-import React from 'react'
-
+import React from "react";
 import { style, layout } from "styles"
 import * as sVar from "constants/styleVariables.js";
+// import { CashBookNoneBtn } from "styles/styled-components/styles";
 
-function CommentDeleteModal({ setClose, onClickHandler, children }) {
+function CashDetailModal({ setClose, onClickHandler, children }) {
+
   return (
     <style.ModalOverlay onClick={setClose}>
       <style.ModalDefault style={{width: "90%", color: "black"}} onClick={(event) => event.stopPropagation()}>
-        {children.map((ment)=><span>{ment}</span>)}
+        {children.map((ment)=> { return <span>{ment}</span> })}
         <layout.FlexCenterRow100 style={{gap:"10px"}}>
         <style.CashBookNoneBtn background={sVar.bookModalYesBtn} onClick={onClickHandler}>예</style.CashBookNoneBtn>
         <style.CashBookNoneBtn background={sVar.bookModalNoBtn} onClick={setClose}>아니오</style.CashBookNoneBtn>
@@ -17,4 +18,6 @@ function CommentDeleteModal({ setClose, onClickHandler, children }) {
   )
 }
 
-export default CommentDeleteModal
+export default CashDetailModal;
+
+// 현재 cashbook쪽과 boardDetail 모두에서 사용하고 있음
