+++ src/features/board/components/CommentDeleteModal.jsx
@@ -1,14 +1,13 @@
-import React from "react";
+import React from 'react'
+
 import { style, layout } from "styles"
 import * as sVar from "constants/styleVariables.js";
-// import { CashBookNoneBtn } from "styles/styled-components/styles";
 
-function CashDetailModal({ setClose, onClickHandler, children }) {
-
+function CommentDeleteModal({ setClose, onClickHandler, children }) {
   return (
     <style.ModalOverlay onClick={setClose}>
       <style.ModalDefault style={{width: "90%", color: "black"}} onClick={(event) => event.stopPropagation()}>
-        {children.map((ment)=> { return <span>{ment}</span> })}
+        {children.map((ment)=><span>{ment}</span>)}
         <layout.FlexCenterRow100 style={{gap:"10px"}}>
         <style.CashBookNoneBtn background={sVar.bookModalYesBtn} onClick={onClickHandler}>예</style.CashBookNoneBtn>
         <style.CashBookNoneBtn background={sVar.bookModalNoBtn} onClick={setClose}>아니오</style.CashBookNoneBtn>
@@ -18,4 +17,6 @@ function CashDetailModal({ setClose, onClickHandler, children }) {
   )
 }
 
-export default CashDetailModal;
+export default CommentDeleteModal
+
+// BoardDetailComment.jsx에서 사용중
