+++ src/features/main/components/MainRecordCard.jsx
