+++ src/features/main/components/chart/Bar.jsx
